# Improve performance with less specialization