~~This is "WIP" because it currently fails [two tests](https://github.com/julia-vscode/SymbolServer.jl/blob/054629cdd0aa8b937b7dda57cbc3692bf88562cd/test/runtests.jl#L42-L45) (just for a handful of `Core.Builltin`and/or `Core.Intrinsic` functions, I think). I'd appreciate some help figuring out how to fix them.~~

~~This takes two main strategies, control over specialization and greater caching, to improve package performance.~~

Several functions have to operate on arguments that have a huge number of different types: any function that takes a function or a type as an argument might, in principle, be specialized, but then used only once or a small handful of times. This marks such arguments with atsign-nospecialize annotations, to prevent them from being recompiled many times over.  Since this package has to use `isa` logic anyway, there is very little advantage in specialization and a big win in terms of compile time for avoiding it.

EDIT: the text below and much of the discussion below is no longer relevant to this PR (those issues will be resolved in a separate PR)

The other major strategy relies on caching. Here there are negatives, because we have to invalidate the cache if the LanguageServer has to process newly-added packages; I'm not sure if that ever happens, and if so where that invalidation should be done.

Of course, the advantage is performance. To major bottlenecks were:
- `Base._methods`: the issue is that two modules might define methods for the same `f` (e.g., `Base.convert` might be extended by multiple packages). However, `Base._methods` looks up methods across the whole session; we might as well save the result.
- `isdefined(m, n)` where `n` came from `allnames`. This was wasteful since it was iterating over all names defined throughout the entire session rather than just the names defined in a specific module.

With all that said, let's get to the numbers. Here was my test:
```julia
using SymbolServer
using Images, Profile, ProfileView  # Revise was also loaded
env = SymbolServer.getenvtree()
@time SymbolServer.symbols(env)

# And then run it a second time in the same session, just for fun (not sure this is relevant)
env = SymbolServer.getenvtree()
@time SymbolServer.symbols(env)
```

And here's the timing results:

| Run # | Time, master | Time, this PR | Gain |
|:----------- | ------------:| -------------:| ----:|
| 1 | 89.9s | 11.3s | 8x |
| 2 | 65.1s |  2.2s | 30x |

As you can see, the gains are pretty decent. I did the timing on top of #159, but I don't think that really changes this test very much.