@ZacLN 2020-05-30
Amazing stuff, I haven't gone through the code in great detail but avoiding specialization seems a clear winner.

> The other major strategy relies on caching. Here there are negatives, because we have to invalidate the cache if the LanguageServer has to process newly-added packages; I'm not sure if that ever happens, and if so where that invalidation should be done.

I'm not entirely clear from my scan of the PR at what level this invalidation happens - is this reference to extended methods from another module? Do you mean that the addition of a package to an environment invalidate caches for all packages?

> isdefined(m, n) where n came from allnames. This was wasteful since it was iterating over all names defined throughout the entire session rather than just the names defined in a specific module.

This was a blunt force attempt to ensure that relevant objects are attached to the `ModuleStore` regardless of whether the name exists within the module, i.e. we want to attach the method defined below within the cache of `A`
```julia
module A 
Base.SomeNonExportedFunctionofBase() =1 
end
```
Similarly, we want `Base` methods for `Core.Int` to be stored within `Base`'s cache (though this doesn't necessitate iterating over all names). There are more efficient approaches to this than the one current taken, but I think this PR stores all methods within the FunctionStore at the parent module? (Is this the link to invalidation mentioned above?) 

I appreciate that none of this is covered in the tests...