@ZacLN 2020-05-30
I think the BuiltIns issue is due to `parentmodule` having different behaviour for non-generics:
```julia
Core.bitcast === Core.Intrinsics.bitcast
parentmodule(Core.bitcast) === Core
parentmodule(Core.Intrinsics.bitcast) === Core
(:bitcast in names(Core, all = true)) === false
(:bitcast in names(Core.Intrinsics, all = true)) === true
```